feat(cangjie_compiler): sync from upstream v1.2.0-alpha.20

+++ include/cangjie/Lex/Token.h
@@ -16,11 +16,11 @@
 #define CANGJIE_LEX_TOKEN_H
 
 #include <cstring>
+#include <cstdint>
 #include <string>
 #include <ostream>
 #include <unordered_map>
 #include <vector>
-#include <cstdint>
 
 #include "cangjie/Basic/Position.h"
 #include "cangjie/Utils/CheckUtils.h"
@@ -184,7 +184,7 @@ struct StringPart {
         this->end = end;
     }
 };
-using TokenVecMap = std::unordered_map<size_t, std::vector<Token>>;
+using TokenVecMap = std::unordered_map<unsigned int, std::vector<Token>>;
 const std::vector<TokenKind>& GetEscapeTokenKinds();
 } // namespace Cangjie
